Moved Conversator to a module folder, moved ConversatorUserData and ConversatorChatMessage classes to the globals
