+++ lib/src/bot.dart
@@ -14,7 +14,7 @@ import 'package:weather/src/modules/panorama_manager.dart';
 import 'package:weather/src/modules/dadjokes_manager.dart';
 import 'package:weather/src/modules/reputation_manager.dart';
 import 'package:weather/src/modules/youtube_manager.dart';
-import 'package:weather/src/modules/conversator_manager.dart';
+import 'package:weather/src/modules/conversator/conversator_manager.dart';
 import 'package:weather/src/modules/general_manager.dart';
 import 'package:weather/src/modules/accordion_poll/accordion_poll_manager.dart';
 import 'package:weather/src/modules/command_statistics/command_statistics_manager.dart';
