+++ lib/src/globals/conversator_chat_message.dart
@@ -1,0 +1,6 @@
+class ConversatorChatMessage {
+  final String message;
+  final bool fromUser;
+
+  ConversatorChatMessage({required this.message, required this.fromUser});
+}
